+++ nix/package.nix
@@ -2,6 +2,7 @@
   lib,
   rustPlatform,
   cacert,
+  installShellFiles,
 }:
 
 rustPlatform.buildRustPackage {
@@ -22,6 +23,15 @@ rustPlatform.buildRustPackage {
   cargoLock.lockFile = ../Cargo.lock;
 
   env.SSL_CERT_FILE = "${cacert}/etc/ssl/certs/ca-bundle.crt";
+
+  postInstall = ''
+    installShellCompletion --cmd vein \
+      --bash <($out/bin/vein completions bash) \
+      --fish <($out/bin/vein completions fish) \
+      --zsh <($out/bin/vein completions zsh)
+  '';
+
+  nativeBuildInputs = [ installShellFiles ];
 
   meta.mainProgram = "vein";
 }